Clamp unit
A Clamp Unit is used to lock the panel position in transitional direction. It supports the panel with fixed Mylar and lock its movement with mobile Mylar, when panels come onto the tool then clamp is in open condition once panel sits down then clamp closes and joining process is done. Once the joining process gets completed, then the cylinder arm opens. Complexity of the unit depends upon the gun reachability and panel area. Given picture is just an example of a simple clamp unit